pyproject.toml
[project]
name = "ai-book-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.124.0",
    "markdown>=3.10",
    "openai-agents>=0.6.2",
    "pypdf>=6.4.0",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.16.1",
    "sentence-transformers>=5.1.2",
    "trafilatura>=2.0.0",
    "uvicorn>=0.38.0",
]
